+++ src/main/java/insane96mcp/iguanatweaksreborn/module/combat/feature/NoKnockback.java
@@ -1,6 +1,7 @@
 package insane96mcp.iguanatweaksreborn.module.combat.feature;
 
 import com.google.common.collect.Multimap;
+import insane96mcp.iguanatweaksreborn.IguanaTweaksReborn;
 import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.iguanatweaksreborn.setup.Strings;
 import insane96mcp.insanelib.base.Feature;
@@ -8,7 +9,9 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
-import insane96mcp.insanelib.util.IdTagMatcher;
+import net.minecraft.core.Registry;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.tags.TagKey;
 import net.minecraft.util.Mth;
 import net.minecraft.world.damagesource.CombatEntry;
 import net.minecraft.world.entity.EquipmentSlot;
@@ -17,23 +20,21 @@ import net.minecraft.world.entity.ai.attributes.Attribute;
 import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.event.entity.living.LivingKnockBackEvent;
 import net.minecraftforge.event.entity.player.AttackEntityEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import net.minecraftforge.registries.ForgeRegistries;
 
-import java.util.ArrayList;
-import java.util.List;
-
-@Label(name = "No Knockback", description = "Player will deal no knockback if attacking with a non-weapon or spamming")
+@Label(name = "No Knockback", description = "Player will deal no knockback if attacking with a non-weapon or spamming.")
 @LoadFeature(module = Modules.Ids.COMBAT)
 public class NoKnockback extends Feature {
 
+	public static final ResourceLocation NO_KNOCKBACK_TAG = new ResourceLocation(IguanaTweaksReborn.MOD_ID + "no_knockback");
+
 	@Config
-	@Label(name = "Custom No Knockback Items", description = "A list of items and tags that should deal no knockback when attacking.")
-	public static List<IdTagMatcher> customNoKnockbackItems = new ArrayList<>();
-	@Config
-	@Label(name = "No Item No Knockback", description = "If true the player will deal no knockback when not using a tool / weapon")
+	@Label(name = "No Item No Knockback", description = "If true the player will deal no knockback when not using an item that doesn't have the attack damage attribute.")
 	public static Boolean noItemNoKnockback = true;
 	@Config(min = 0d, max = 1d)
 	@Label(name = "Attack Cooldown No Knockback", description = "When the attack cooldown is below this percentage the player will deal no knockback. (Between 0 and 1, where 1 is the attack fully charged)")
@@ -66,16 +67,13 @@ public class NoKnockback extends Feature {
 		if (combatEntry == null || !(combatEntry.getSource().getDirectEntity() instanceof Player))
 			return;
 		ItemStack itemStack = player.getMainHandItem();
-		boolean isInList = false;
-		for (IdTagMatcher idTagMatcher : customNoKnockbackItems) {
-			if (idTagMatcher.matchesItem(itemStack.getItem(), null)) {
-				isInList = true;
-				break;
-			}
-		}
+
+		TagKey<Item> tagKey = TagKey.create(Registry.ITEM_REGISTRY, NO_KNOCKBACK_TAG);
+		boolean isInTag = ForgeRegistries.ITEMS.tags().getTag(tagKey).contains(itemStack.getItem());
+
 		boolean preventKnockback = false;
 		Multimap<Attribute, AttributeModifier> attributeModifiers = itemStack.getAttributeModifiers(EquipmentSlot.MAINHAND);
-		if ((!attributeModifiers.containsKey(Attributes.ATTACK_DAMAGE) && noItemNoKnockback) || isInList) {
+		if ((!attributeModifiers.containsKey(Attributes.ATTACK_DAMAGE) && noItemNoKnockback) || isInTag) {
 			preventKnockback = true;
 		}
 		int ticksSinceLastSwing = player.getPersistentData().getInt(Strings.Tags.TIME_SINCE_LAST_SWING);