+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Upcoming
+* [Moved many blacklists to data tags](https://github.com/Insane96/IguanaTweaksReborn/wiki/%5B1.19-and-up%5D-Tags) and added some new ones
+* Fixed some strings not begin translatable
+* Fixed tiredness multiplier not working
+
 ## Alpha 2.13.1
 * Requires InsaneLib 1.7.1
 * Added back Client module